install node js with dependences and create api admin
