+++ .env.dev
@@ -1,2 +1,2 @@
-DATABASE_URL="mysql://root:@localhost:3306/myhotel?schema=public"
-PORT=3306
+PORT=2000
+DATABASE_URL=postgresql://hilmi:12November2000@127.0.0.1:5432/db_express
