+++ src/index.js
@@ -2,6 +2,7 @@ const express = require("express");
 const dotenv = require("dotenv");
 const app = express();
 const cors = require("cors");
+const router = require("./routes/router");
 dotenv.config();
 const PORT = process.env.PORT;
 app.use(express.json());
@@ -11,16 +12,10 @@ app.use(
   })
 );
 
-const userController = require("./admin/admin.controller");
-const roomController = require("./room/room.controller");
-const floorController = require("./floor/floor.controller");
-
 app.get("/api", (req, res) => {
   res.send("Hello");
 });
-app.use("/users", userController);
-app.use("/rooms", roomController);
-app.use("/floor", floorController);
+app.use("/", router);
 
 app.listen(PORT, () => {
   console.log("Express API running in Port: " + PORT);
