+++ src/module/admin/admin.service.js
@@ -1,4 +1,3 @@
-const prisma = require("../db/index");
 const {
   findAllUser,
   findUser,
