+++ src/module/room/room.controller.js
@@ -1,0 +1,12 @@
+const express = require("express");
+
+const { allUsers } = require("./room.service");
+const router = express.Router();
+
+router.get("/", async (req, res) => {
+  const users = await allUsers();
+
+  res.send(users);
+});
+
+module.exports = router;
